+++ examples/basic_tutorials/tutorial_mnist_mlp_dynamic.py
@@ -46,7 +46,7 @@ MLP = CustomModel()
 n_epoch = 500
 batch_size = 500
 print_freq = 5
-train_weights = MLP.weights
+train_weights = MLP.trainable_weights
 optimizer = tf.optimizers.Adam(learning_rate=0.0001)
 
 ## the following code can help you understand SGD deeply