+++ examples/basic_tutorials/tutorial_mnist_mlp_static.py
@@ -37,7 +37,7 @@ MLP = get_model([None, 784])
 n_epoch = 500
 batch_size = 500
 print_freq = 5
-train_weights = MLP.weights
+train_weights = MLP.trainable_weights
 optimizer = tf.optimizers.Adam(lr=0.0001)
 
 ## the following code can help you understand SGD deeply