+++ tensorlayer/layers/normalization.py
@@ -258,8 +258,12 @@ class BatchNorm(Layer):
         if self.gamma_init:
             self.gamma = self._get_weights("gamma", shape=params_shape, init=self.gamma_init)
 
-        self.moving_mean = self._get_weights("moving_mean", shape=params_shape, init=self.moving_mean_init)
-        self.moving_var = self._get_weights("moving_var", shape=params_shape, init=self.moving_var_init)
+        self.moving_mean = self._get_weights(
+            "moving_mean", shape=params_shape, init=self.moving_mean_init, trainable=False
+        )
+        self.moving_var = self._get_weights(
+            "moving_var", shape=params_shape, init=self.moving_var_init, trainable=False
+        )
 
     def forward(self, inputs):
         mean, var = tf.nn.moments(inputs, self.axes, keepdims=True)