+++ tests/layers/test_layers_recurrent.py
@@ -62,8 +62,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y, final_state = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -132,8 +132,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -167,8 +167,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -209,8 +209,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -236,8 +236,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y, final_h, final_c = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -263,8 +263,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y, final_h = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -296,8 +296,8 @@ class Layer_RNN_Test(CustomTestCase):
             self.assertEqual(
                 r.get_shape().as_list(), [self.batch_size * self.num_steps, self.hidden_size + self.hidden_size + 1]
             )
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -330,8 +330,8 @@ class Layer_RNN_Test(CustomTestCase):
             self.assertEqual(
                 r.get_shape().as_list(), [self.batch_size, self.num_steps, self.hidden_size + self.hidden_size + 1]
             )
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -367,8 +367,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -401,8 +401,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))
@@ -438,8 +438,8 @@ class Layer_RNN_Test(CustomTestCase):
                 pred_y = rnn_model(self.data_x)
                 loss = tl.cost.mean_squared_error(pred_y, self.data_y2)
 
-            gradients = tape.gradient(loss, rnn_model.weights)
-            optimizer.apply_gradients(zip(gradients, rnn_model.weights))
+            gradients = tape.gradient(loss, rnn_model.trainable_weights)
+            optimizer.apply_gradients(zip(gradients, rnn_model.trainable_weights))
 
             if (epoch + 1) % 10 == 0:
                 print("epoch %d, loss %f" % (epoch, loss))