+++ CHANGELOG.md
@@ -76,6 +76,8 @@ To release a new version, please update the changelog as followed:
 ### Changed
 - remove `tl.layers.initialize_global_variables(sess)` (PR #931)
 - update `tutorial_generate_text.py`, `tutorial_ptb_lstm.py`. remove `tutorial_ptb_lstm_state_is_tuple.py` (PR #958)
+- change `tl.layers.core`, `tl.models.core` (PR #966)
+- change `weights` into `all_weights`, `trainable_weights`, `nontrainable_weights`
 
 ### Dependencies Update
 - nltk>=3.3,<3.4 => nltk>=3.3,<3.5 (PR #892)